+++ storage.go
@@ -4,23 +4,21 @@ import (
 	"fmt"
 
 	"github.com/Luzifer/ots/pkg/storage"
-	"github.com/Luzifer/ots/pkg/storage/memory"
-	"github.com/Luzifer/ots/pkg/storage/redis"
+	"github.com/Luzifer/ots/pkg/storage/factory"
 )
 
 func getStorageByType(t string) (storage.Storage, error) {
 	switch t {
 	case "mem":
-		return memory.New(), nil
+		return factory.CreateStorageEngine("memory://")
 
 	case "redis":
-		s, err := redis.New()
-		if err != nil {
-			return s, fmt.Errorf("creating redis storage: %w", err)
-		}
-		return s, nil
-
-	default:
-		return nil, fmt.Errorf("storage type %q not found", t)
+		return factory.CreateStorageEngine("redis://")
 	}
+
+	s, err := factory.CreateStorageEngine(t)
+	if err != nil {
+		return nil, fmt.Errorf("creating storage engine for %q: %w", t, err)
+	}
+	return s, nil
 }